Move all relavant user related to the UserViewModel
